+++ view/generic_view.py
@@ -3,27 +3,6 @@ from getpass import getpass
 
 class View:
     @staticmethod
-    def quit_option():
-        print("   Q) pour quitter")
-
-    @staticmethod
-    def is_quitting(choice: str) -> bool:
-        return choice.lower() == "q"
-
-    @staticmethod
-    def valid_choices(max_range: int) -> list:
-        quit_choices: list = ["q", "Q"]
-        valid_choices_list = [str(i) for i in range(1, max_range + 1)]
-
-        return valid_choices_list + quit_choices
-
-    @staticmethod
-    def unknown_option():
-        print(
-            "\nCette option n'existe malheureusement pas,"
-            " veuillez sélectionner une commande valide parmi la liste\n")
-
-    @staticmethod
     def connection() -> tuple[str, str]:
         print("Pour vous connecter, veuillez renseigner votre nom d'utilisateur:")
         username = input("")
@@ -54,3 +33,34 @@ class View:
     def is_that_you(username: str) -> bool:
         print(f"Êtes-vous bien {username} ?")
         return View.yes_or_no_choice()
+
+    @staticmethod
+    def quit_option():
+        print("   Q) pour quitter")
+
+    @staticmethod
+    def is_quitting(choice: str) -> bool:
+        return choice.lower() == "q"
+
+    @staticmethod
+    def valid_choices(max_range: int) -> list:
+        quit_choices: list = ["q", "Q"]
+        valid_choices_list = [str(i) for i in range(1, max_range + 1)]
+
+        return valid_choices_list + quit_choices
+
+    @staticmethod
+    def unknown_option():
+        print(
+            "\nCette option n'existe malheureusement pas,"
+            " veuillez sélectionner une commande valide parmi la liste\n")
+
+    @staticmethod
+    def print_choices(choices_list: list) -> None:
+        [print(f"   {index + 1}) {choices_list[index]}") for index in range(0, len(choices_list))]
+
+    @classmethod
+    def is_choice_valid(cls, choice: str, max_range: int) -> bool:
+        return any(choice == valid for valid in cls.valid_choices(max_range))
+
+
